LarCongruence.cellCongruenceAA_OPT — Function

cellCongruenceAA_OPT(Delta, inclasses)

Valuta la congruenza tra celle per una parte di topologia Delta utilizzando le classi inclasses, cioè la mappa che identifica per ciascun nuovo vertice calcolato i vecchi vertici di partenza, oppure la mappa che identifica per ciascun nuovo bordo calcolato i vecchi bordi di partenza.

La funzione restituisce: - la matrice che rappresenta i nuovi bordi calcolati e la mappa che li lega ai bordi di partenza se in input viene passato un insieme di vertici e la mappa che li lega ai vertici di partenza. - la matrice che rappresenta le nuove facce calcolate e la mappa che le lega alle facce di partenza se in input viene passato un insieme di bordi e la mappa che li lega ai bordi di partenza.

Argomenti

Delta::SparseMatrixCSC{Int8, Int64}

inclasses::Vector{Any}

Return

(FEnew, outclasses)::Tuple{ Vector{Vector{Int64}}, Vector{Vector{Int64}}}

LarCongruence.chainCongruenceAA_OPT — Function

chainCongruenceAA_OPT(W, T)

Calcola la congruenza della geometria W e coerentemente trasforma la topologia T La funzione restituisce tra matrici di incidenza/adiacenza: - V (il nuovo insieme di vertici calcolato con vertCongruenceAAOPT) - EV (matrice di adiacenza bordi/vertici, calcolata con cellCongruenceOPT utilizzando vclasses) - FE (matrice di adiacenza facce/bordi, calcolata con cellCongruence_OPT utilizzando eclasses)

Argomenti

W::Lar.Points

T::Vector{SparseMatrixCSC{Int8, Int64}}

Return

(V, EV, FE)::Tuple{Lar.Points, Vector{Vector{Int64}}, Vector{Vector{Int64}}}