LarCongruence.cellCongruence_OPT — Function

cellCongruence_OPT(Delta, classes)

Valuta ed esegue la congruenza tra celle per una geometria (in graphBLAS) Delta utilizzando le classi `classes', cioè la mappa che identifica per ciascun nuovo vertice calcolato identifica i vecchi vertici di partenza.

Argomenti

Delta::GrB_Matrix{T}

classes::Array{Array{Int,1},1}

Return

(res, eclasses)::Tuple{Matrix{Float64}, Vector{GrB_Matrix{Int8}}}

LarCongruence.chainCongruenceGB_OPT — Function

chainCongruenceGB_OPT(W, Top)

Calcola la congruenza della geometria W e coerentemente trasforma la topologia Top

Argomenti

W::Lar.Points

Top::Vector{GrB_Matrix{Int8}}

Return

(V, Topn)::Tuple{Lar.Points, TnArray{Lar.ChainOp}}