LarCongruence.cellCongruenceSM_OPT — Function

cellCongruenceSM_OPT(cop, lo_cls, lo_sign)

Valuta la congruenza tra celle per una cocatena cop utilizzando le classi lo_cls, cioè la mappa che identifica per ciascun nuovo vertice calcolato identifica i vecchi vertici di partenza.

La funzione determina il nuovo operatore di cocatena costruito a partire da cop dove viene eseguito il merge delle celle di ordine inferiore secondo la mappa lo_cls. lo_sign specifica se una cella debba essere considerata nel verso opposto.

Se il parametro opzionale imp vale true, allora le imprecizioni vengono tenute in considerazione nel senso che le celle di ordine inferiore potrebbero collidere. Il parametro d rappresenta, quindi l'ordine della cella (corrisponde anche al numero minimo di celle di ordine inferiore di cui la cella corrente deve essere costituita).

Argomenti

cop::Lar.ChainOp

lo_cls::Array{Array{Int,1},1}

lo_sign::Array{Array{Int8,1},1}

imp = false

d = 0

Return

(nrows, ho_cls, ho_sign)::Tuple{ Lar.ChainOp, Array{Array{Int,1},1}, Array{Array{Int8,1},1} }

LarCongruence.chainCongruenceSM_OPT — Function

chainCongruenceSM_OPT(G, T)

Calcola la congruenza della geometria G e coerentemente trasforma la topologia T

Argomenti

G::Lar.Points

T::Array{Lar.ChainOp}

imp=false

ϵ=1e-6

Return

(G, Tn)::Tuple{Lar.Points, TnArray{Lar.ChainOp}}