Vertices Congruence

LarCongruence.vertCongruenceOptimized — Function

vertCongruenceOptimized(V)

Valuta ed esegue la congruenza di vertici per punti 3D $V ∈ ℳ(3,n)$.

La funzione determina i punti di V più vicini di ϵ e costruisce un nuovo insieme di vertici costituito dal rappresentativo per ciascun cluster di punti. Per fare ciò si utilizza l'algoritmo KDTree fornito dal package NearestNeighbors.

Il metodo restituisce:

il nuovo insieme di vertici

una mappa che, per ogni nuovo vertice, identifica i vecchi vertici nell'insieme di partenza

Argomenti

'V::Lar.Array{Float64,2}'

'ϵ=1e-6'

Return

(V, Vcls)::Tuple{Lar.Points, Array{Array{Int,1},1}}

err = 1e-8
V = [
    0.0  err  0.0 -err  0.0  0.0
    0.0  0.0  err  0.0 -err  0.0
]

LC.vertCongruenceOptimized(V[:, 1:5])

LC.vertCongruenceOptimized(V[:, 2:6])

Variante per implementazione AA

LarCongruence.vertCongruenceAA_OPT — Function

vertCongruenceAA_OPT(W)

Valuta ed esegue la congruenza di vertici per punti 3D $W ∈ ℳ(3,n)$.

La funzione determina i punti di W più vicini di err (definito all'interno della funzione, valore pari a 10^-6) e costruisce un nuovo insieme di vertici costituito dal rappresentativo per ciascun cluster di punti. Per fare ciò si utilizza l'algoritmo KDTree fornito dal package NearestNeighbors.

Il metodo restituisce:

il nuovo insieme di vertici

una mappa che, per ogni nuovo vertice, identifica i vecchi vertici nell'insieme di partenza

Argomenti

'V::Lar.Array{Float64,2}'

Return

(V, vclasses)::Tuple{Lar.Points, Array{Array{Int,1},1}}